pi-extension: simplify per review (drop phase2 module, collapse reload state, extract attribution)
Behavior-preserving structural cleanup from a maintainability review:
- Delete the handlers/phase2.ts grab-bag: provider request/response
  enrichment moves to llm.ts (reusing its current-LLM lookup, removing a
  duplicated helper); the compaction span moves to session.ts. index.ts no
  longer registers it.
- Collapse the redundant resumeParent/resumeFrom session state into a single
  resumeFrom field; the remote-parent context is built on demand.
- Extract source attribution into attribution.sessionAttributes(), applied in
  one loop instead of repeated inline setAttr calls.

No behavior change; tsc clean; 54 tests pass.
